+++ CHANGELOG.md
@@ -1,6 +1,7 @@
  == CHANGELOG ==
 
 * Screenshot supersampling is now saved as a setting.
+* Skyrim Special Edition particle data is updated by spells that modify BSTriShape geometry.
 * Fixed the "Extract Resource Files" spell not extracting compiled Starfield materials.
 
 #### NifSkope-2.0.dev11-20250414